PersonTile moved to common folder
