--- src/common/Tiles/PersonTile/TilesContainer/PersonsTile/index.js
@@ -1,12 +1,0 @@
-import { Tile, Image, Name } from "./styled";
-
-const PersonTile = ({ image, name }) => {
-  return (
-    <Tile>
-      <Image src={image} alt="Person Image"></Image>
-      <Name>{name}</Name>
-    </Tile>
-  );
-};
-
-export default PersonTile;
