--- src/common/Tiles/PersonTile/TilesContainer/PersonsTile/styled.js
@@ -1,31 +1,0 @@
-import styled from "styled-components";
-
-export const Tile = styled.div`
-  width: 208px;
-  height: 339px;
-  border-radius: 5px;
-  background-color: ${({ theme }) => theme.color.white};
-  box-shadow: 0px 4px 12px 0px #bac7d550;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-export const Image = styled.img`
-  margin-top: 16px;
-  margin-bottom: 12px;
-  width: 176px;
-  height: 231x;
-  border-radius: 5px;
-  padding-bottom: 0;
-  text-align: center;
-`;
-
-export const Name = styled.h1`
-  font-size: 22px;
-  font-weight: 500;
-  line-height: 28.6px;
-  text-align: center;
-  color: ${({ theme }) => theme.color.woodsmoke};
-  margin: 0 16px;
-`;
