--- src/common/Tiles/PersonTile/TilesContainer/PersonsTitle/index.js
@@ -1,7 +1,0 @@
-import { StyledPersonsTitle } from "./styled";
-
-const PersonsTitle = ({ title }) => {
-  return <StyledPersonsTitle>{title}</StyledPersonsTitle>;
-};
-
-export default PersonsTitle;
