--- src/common/Tiles/PersonTile/TilesContainer/PersonsTitle/styled.js
@@ -1,9 +1,0 @@
-import styled from "styled-components";
-
-export const StyledPersonsTitle = styled.h1`
-  font-size: 36px;
-  font-weight: 600;
-  line-height: 43.2px;
-  margin-top: 56px;
-  margin-bottom: 24px;
-`;
