--- src/common/Tiles/PersonTile/TilesContainer/index.js
@@ -1,13 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-  max-width: 1368px;
-  margin: 0 auto;
-  margin-bottom: 56px;
-`;
-
-export const Tiles = styled.div`
-  display: grid;
-  grid-gap: 24px;
-  grid-template-columns: 1fr 1fr 1fr 1fr 1fr 1fr;
-`;
