+++ src/common/Tiles/PersonTile/index.js
