+++ src/common/Tiles/PersonTile/styled.js
